Reorganize io module
